Move queries to graphql folder
