+++ frontend/src/graphql/queries.ts
